Throw error at inconsistent UVISOR_PRESENT settings
The UVISOR_PRESENT variable is set by the uvisor-lib CMakeLists.txt
file for its internal use, and it's based on the actual list of officially
supported platforms.

Targets, on the other hand, need to define a yotta config variable:
YOTTA_CFG_UVISOR_PRESENT
to confirm that uVisor has been ported and it is now supported on that
platform.

This commit adds a sanity check on the consistency between
CMakeLists.txt::UVISOR_PRESENT and YOTTA_CFG_UVISOR_PRESENT. It will prevent
copy-paste mistakes that are likely to be made when using an existing config
file as a starting point for a new target.