Add explicit vIRQ_SetVector API to override.h (mbed)
In this previous commit:
https://github.com/ARMmbed/uvisor/commit/f83a82d881207f28ce01bd34e67d8df9dd0b9822
("vIRQ_SetVector turned into an actual function")
vIRQ_SetVector has been turned in an actual API instead of a preprocessor macro
redefinition.

This commit mirrors that change to the override.h file, which is used in mbed
to override NVIC symbols when one does not want to replace them manually in the
source code.

An example use case:
https://github.com/ARMmbed/mbed-hal-st-stm32f4/blob/master/source/hal_tick.c#L36.