Allow owning system IRQs with UVISOR_DISABLED
Code written to work with uVisor would never touch system IRQs because
they are restricted. So if `__uvisor_mode == UVISOR_DISABLED` it was
natural to only allow user IRQs in interrupt manipulation (`vIRQ_*`
APIs).

The problem is that `__uvisor_mode == UVISOR_DISABLED` also when code
was not written to work with uVisor but it runs on a supported platform
(by default a supported platform has `YOTTA_CFG_UVISOR_PRESENT == 1` but
`__uvisor_mode == UVISOR_DISABLED`).

This commit allows `vIRQ_*` APIs to work with system IRQs when the
uVisor is disabled.

Fixes: #36 ("uvisor will not allow SysTick_IRQn to be reconfigured")
Fixes: ARMmbed/uvisor#e44d0c8 ("Add IRQ context switch fallback for disabled.cpp")

This commit mirrors:
ARMmbed/uvisor#b183d4a ("Allow owning system IRQs with UVISOR_DISABLED")