Workaround for bug in memory access macros
This is a temporary workaround - #173 has been created to keep track of
the original bug.

If optimizations are turned on, the r0 register, in which the address
for the memory access is stored, might contain garbage. The signature
for the `uvisor_{read, write}{8, 16, 32}` APIs only specified a volatile
data-pointed-to, but not a volatile pointer.

Mirrors:
ARMmbed/uvisor@5177e9e ("Apply workaround for bug in memory access macros")
ARMmbed/uvisor@b6f0e50 ("Swap uintX_t and volatile in type definitions")